DBTestBase keeps an open connection to H2 to preserve in-memory DB.
This change fixes a problem with in-memory H2 databases introduced
with r397.  It maintains an open connection to the database for the
duration of a test, because H2 will delete an in-memory DB when the
last connection to it is closed.

Many of the tests were spuriously failing because the in-memory test
databases were getting deleted when there was no outstanding
connection to them.  This would often happen before an individual test
finished running.

Code Review:
http://codereview.appspot.com/6495116