Added a MockClient implementation to replace the iBATIS-backed DBClient. The MockClient is in the prod tree so that you can use it on a deployed database connector instance. The client implementation can be swapped out in connectorInstance.xml. For validation, you need to configure the connector against a database, but you can use an in-memory H2 database that is conjured out of the air on the config form itself.
The connectorDefaults.xml changes are compatible with existing
connectorInstance.xml files.

The DBContext is no longer available from the DBClient, since the
DBClient is obtained from the DBContext. This move changes when
exceptions occur, which led to some extra refactoring of the tests
when catch blocks had to be removed.

Code review: http://codereview.appspot.com/6597045/