Hack to use zero-padded primary keys for an X build based on 2.8.4. This is intended to get a customer past these problems while we fix this for the general cases.
This hack assumes that the primary key is a single 4-byte
integer-valued column, and instead of Base64-encoding it, it simply
zero-pads it to 10 digits.

I also included a fix from r397 that lets me test this against H2.

As a note, customers could easily zero-pad their keys, but they can't
avoid the problem with the Base64-encoding not preserving the sort
order.

Code review: https://codereview.appspot.com/7027052