Modified numeric docid hack to support any Number subclasses.
The customer column type turned out to be NUMBER(22), which is
returned as a java.math.BigDecimal. For this hack, I just switched to
assume it's any Number (which adds support for Float, Double, and
BigDecimal), and call longValue on it before formatting.

Bumped version on the branch to 2.8.5-X2.

Code review: https://codereview.appspot.com/7066057/