Fix Bug 7732455, part 2: Preserve Sort Order in Encoded DocIds - include types
This change enhances the robustness of Comparable DocumentSnapshots.
This change encodes the types of the primary key values into the docid.
This allows us to distinguish between a number and a string that looks
like a number, for instance.  It also allows us to distinguish between
null values and empty string.  Preserving the types also lays the 
groundwork for reconstructing the values as their proper type when we
fix authz query construction.
See also:
b/7928861: Avoid unordered snapshots in the snapshot files.

Code review:
http://codereview.appspot.com/7069070