Fixed a Spring XML syntax error in r458 and a MyBatis regression in r438 that broke authorization queries.
The collator property uses util:constant to access JDBC constants, but
that requires using an XML schema instead of the DTD. For simplicity,
I switched to a FieldRetrievingFactoryBean to get the values.

The upgrade to MyBatis 3 switched the query syntax parameter from
$docIds$ to ${docIds}, which Spring interprets as a placeholder. I
added a property value with an escaped $ to pass ${docIds} through.

The MockDBConnectorFactory did not actually instantiate the connector
instance, which would have identified both of these errors. I mimicked
the code in InstanceInfo, with some local simplifications. I put in a
TODO and there's also an FR to extract the CM code for use in
connector tests.

Other changes:

* Cleaned up the constants in TestUtils. Moved the Spring XML
  constants to MockDBConnectorFactory, the only place they were used.
  Deleted unused constants for SQL data files. Made the rest private.

* Removed the SqlFunction from DBConnectorTypeTest in favor of the one
  in DBClient. This was a TODO from r451 that was waiting for r455.

Code review: https://codereview.appspot.com/10131043/