Fix bug 7301515 redux: Reduce the memory footprint of LOBs in iBATIS.
The earlier fix in r446 fails with SQL Server LOB types, e.g.,
longvarchar(max), as well as the TEXT and IMAGE types, and Oracle's
LONG type. These are reported as LONGVARCHAR or LONGVARBINARY, rather
than CLOB or BLOB.

Global type handlers must be registered by the concrete Java type
of the column values. Since we don't know the concrete types, we
must register the type handlers directly by column name. We need to
know the column types, and we could easily get that in
DBConnectorType, but we don't currently have access to the column
types in DBClient. Instead, we register this generic type handler,
and use the strategy pattern to pick an implementation based on the
column type on the fly.

Other changes:

* Added logging for diffing serializations, type handler loading, and
  LOB column types.

* Required Java 6 by using BLOB.free(), CLOB.free(), and
  String.getBytes(Charset).

Unit tests to follow in a separate CL.

Code review: https://codereview.appspot.com/7329048/