Improve document content options on the configuration form, part two.
Added tests to describe the behavior when one of the three original
radio buttons is selected, but the corresponding fields are left
empty. The "notExt" or Stylesheet option is implied, but not displayed
as such. In a later part, such a configuration will be rejected if
submitted, and the implied configuration will be displayed on the
populated form.

Other changes:

* Refactor ValidateUtil.validate to make the structure easier to see
  and change.

* Use the DBConnectorType field constants in DBConnectorTypeTest,
  making a few public (chosen over package access for consistency) as
  required.

* Create a new config Map for each test, rather than relying on
  overwriting all of the desired properties.

Code review: https://codereview.appspot.com/14600044/