Improve document content options on the configuration form, part five.
Fix the initial display of sometimes required fields. The required
asterisks were not displayed for these fields, but only made visible
by JavaScript when you selected the corresponding radio buttons. This
is similar to part four, which made the same change for disabling the
fields.

Added checks for the sometimes required fields to the tests.

Code review: https://codereview.appspot.com/15230043